Implement packaging for focus and update assertions
