+++ src/main/java/wellnus/WellNus.java
@@ -2,7 +2,7 @@ package wellnus;
 
 import wellnus.common.MainManager;
 import wellnus.exception.BadCommandException;
-import wellnus.focus.FocusManager;
+import wellnus.focus.feature.FocusManager;
 import wellnus.manager.Manager;
 import wellnus.ui.TextUi;
