+++ src/main/java/wellnus/focus/command/CheckCommand.java
@@ -1,29 +1,35 @@
-package wellnus.focus;
+package wellnus.focus.command;
 
 import wellnus.command.Command;
 import wellnus.exception.BadCommandException;
+import wellnus.focus.feature.FocusManager;
+import wellnus.focus.feature.Session;
 import wellnus.ui.TextUi;
 
 import java.util.HashMap;
 
-public class PauseCommand extends Command {
-    private static final String COMMAND_KEYWORD = "pause";
+/**
+ * Represents a command to check the time left in the current session.
+ */
+public class CheckCommand extends Command {
+
+    private static final String COMMAND_KEYWORD = "check";
     private static final int COMMAND_NUM_OF_ARGUMENTS = 1;
-    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'pause'";
+    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'check'";
     private static final String NO_ADDITIONAL_MESSAGE = "";
-    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be pause.";
-    private static final String PAUSE_OUTPUT = "Timer paused at: ";
+    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be check.";
+    private final String CHECK_OUTPUT = "Time left: ";
     private final Session session;
     private final TextUi textUi;
 
     /**
-     * Constructor for PauseCommand object.
+     * Constructs a CheckCommand object.
      * Session in FocusManager is passed into this class.
      *
      * @param arguments Argument-Payload Hashmap generated by CommandParser
      * @param session   Session object which is an arraylist of Countdowns
      */
-    public PauseCommand(HashMap<String, String> arguments, Session session) {
+    public CheckCommand(HashMap<String, String> arguments, Session session) {
         super(arguments);
         this.session = session;
         this.textUi = new TextUi();
@@ -53,8 +59,8 @@ public class PauseCommand extends Command {
     }
 
     /**
-     * Pause the current countdown.
-     * Prints the time left in the current countdown for user to view.
+     * Checks the current time left in the current countdown.
+     * Prints the time left in the current countdown.
      */
     @Override
     public void execute() {
@@ -64,10 +70,9 @@ public class PauseCommand extends Command {
             textUi.printErrorFor(badCommandException, NO_ADDITIONAL_MESSAGE);
             return;
         }
-        session.getSession().get(session.getCurrentCountdownIndex()).setPause();
         int minutes = session.getSession().get(session.getCurrentCountdownIndex()).getMinutes();
         int seconds = session.getSession().get(session.getCurrentCountdownIndex()).getSeconds();
-        textUi.printOutputMessage(PAUSE_OUTPUT + String.format("%d:%d", minutes, seconds));
+        textUi.printOutputMessage(CHECK_OUTPUT + String.format("%d:%d", minutes, seconds));
     }
 
     /**
