+++ src/main/java/wellnus/focus/command/HomeCommand.java
@@ -1,9 +1,11 @@
-package wellnus.focus;
+package wellnus.focus.command;
 
 import java.util.HashMap;
 
 import wellnus.command.Command;
 import wellnus.exception.BadCommandException;
+import wellnus.focus.feature.FocusManager;
+import wellnus.focus.feature.Session;
 import wellnus.ui.TextUi;
 
 /**
