+++ src/main/java/wellnus/focus/command/PauseCommand.java
@@ -1,32 +1,31 @@
-package wellnus.focus;
+package wellnus.focus.command;
 
 import wellnus.command.Command;
 import wellnus.exception.BadCommandException;
-import wellnus.exception.WellNusException;
+import wellnus.focus.feature.FocusManager;
+import wellnus.focus.feature.Session;
 import wellnus.ui.TextUi;
-
 
 import java.util.HashMap;
 
-public class ResumeCommand extends Command {
-
-    private static final String COMMAND_KEYWORD = "resume";
+public class PauseCommand extends Command {
+    private static final String COMMAND_KEYWORD = "pause";
     private static final int COMMAND_NUM_OF_ARGUMENTS = 1;
-    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'resume'";
+    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'pause'";
     private static final String NO_ADDITIONAL_MESSAGE = "";
-    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be resume.";
-    private final static String RESUME_OUTPUT = "Timer resumed at: ";
+    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be pause.";
+    private static final String PAUSE_OUTPUT = "Timer paused at: ";
     private final Session session;
     private final TextUi textUi;
 
     /**
-     * Constructs a ResumeCommand object.
+     * Constructor for PauseCommand object.
      * Session in FocusManager is passed into this class.
      *
      * @param arguments Argument-Payload Hashmap generated by CommandParser
      * @param session   Session object which is an arraylist of Countdowns
      */
-    public ResumeCommand(HashMap<String, String> arguments, Session session) {
+    public PauseCommand(HashMap<String, String> arguments, Session session) {
         super(arguments);
         this.session = session;
         this.textUi = new TextUi();
@@ -55,10 +54,9 @@ public class ResumeCommand extends Command {
         return FocusManager.FEATURE_NAME;
     }
 
-
     /**
-     * Resumes the current countdown.
-     * Prints the time left to be completed in the current countdown.
+     * Pause the current countdown.
+     * Prints the time left in the current countdown for user to view.
      */
     @Override
     public void execute() {
@@ -68,16 +66,20 @@ public class ResumeCommand extends Command {
             textUi.printErrorFor(badCommandException, NO_ADDITIONAL_MESSAGE);
             return;
         }
+        session.getSession().get(session.getCurrentCountdownIndex()).setPause();
         int minutes = session.getSession().get(session.getCurrentCountdownIndex()).getMinutes();
         int seconds = session.getSession().get(session.getCurrentCountdownIndex()).getSeconds();
-        textUi.printOutputMessage(RESUME_OUTPUT + String.format("%d:%d", minutes, seconds));
-        session.getSession().get(session.getCurrentCountdownIndex()).setStart();
+        textUi.printOutputMessage(PAUSE_OUTPUT + String.format("%d:%d", minutes, seconds));
     }
 
     /**
      * Validate the arguments and payloads from a commandMap generated by CommandParser.
-     * User input is valid if no arguments are thrown.
+     * User input is valid if no exceptions are thrown.
+     *
+     * @param arguments Argument-Payload map generated by CommandParser
+     * @throws BadCommandException If the arguments have any issues
      */
+    @Override
     public void validateCommand(HashMap<String, String> arguments) throws BadCommandException {
         if (arguments.size() != COMMAND_NUM_OF_ARGUMENTS) {
             throw new BadCommandException(COMMAND_INVALID_ARGUMENTS_MESSAGE);
