+++ src/main/java/wellnus/focus/command/ResumeCommand.java
@@ -1,31 +1,33 @@
-package wellnus.focus;
+package wellnus.focus.command;
 
 import wellnus.command.Command;
 import wellnus.exception.BadCommandException;
-import wellnus.exception.WellNusException;
+import wellnus.focus.feature.FocusManager;
+import wellnus.focus.feature.Session;
 import wellnus.ui.TextUi;
+
 
 import java.util.HashMap;
 
-public class StopCommand extends Command {
-    private static final String COMMAND_KEYWORD = "stop";
-    private static final int COMMAND_NUM_OF_ARGUMENTS = 1;
-    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'stop'";
-    private static final String NO_ADDITIONAL_MESSAGE = "";
-    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be stop.";
-    private static final String STOP_MESSAGE = "Your focus session has ended.";
+public class ResumeCommand extends Command {
 
+    private static final String COMMAND_KEYWORD = "resume";
+    private static final int COMMAND_NUM_OF_ARGUMENTS = 1;
+    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'resume'";
+    private static final String NO_ADDITIONAL_MESSAGE = "";
+    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be resume.";
+    private final static String RESUME_OUTPUT = "Timer resumed at: ";
     private final Session session;
     private final TextUi textUi;
 
     /**
-     * Constructs a StopCommand object.
+     * Constructs a ResumeCommand object.
      * Session in FocusManager is passed into this class.
      *
      * @param arguments Argument-Payload Hashmap generated by CommandParser
      * @param session   Session object which is an arraylist of Countdowns
      */
-    public StopCommand(HashMap<String, String> arguments, Session session) {
+    public ResumeCommand(HashMap<String, String> arguments, Session session) {
         super(arguments);
         this.session = session;
         this.textUi = new TextUi();
@@ -56,9 +58,8 @@ public class StopCommand extends Command {
 
 
     /**
-     * Prints message to indicate session has ended.
-     * Stops the session.
-     * Resets the current countdown index.
+     * Resumes the current countdown.
+     * Prints the time left to be completed in the current countdown.
      */
     @Override
     public void execute() {
@@ -68,19 +69,17 @@ public class StopCommand extends Command {
             textUi.printErrorFor(badCommandException, NO_ADDITIONAL_MESSAGE);
             return;
         }
-        textUi.printOutputMessage(STOP_MESSAGE);
-        session.getSession().get(session.getCurrentCountdownIndex()).setStop();
-        session.resetCurrentCountdownIndex();
+        assert super.getArguments().containsKey(COMMAND_KEYWORD) : COMMAND_KEYWORD_ASSERTION;
+        int minutes = session.getSession().get(session.getCurrentCountdownIndex()).getMinutes();
+        int seconds = session.getSession().get(session.getCurrentCountdownIndex()).getSeconds();
+        textUi.printOutputMessage(RESUME_OUTPUT + String.format("%d:%d", minutes, seconds));
+        session.getSession().get(session.getCurrentCountdownIndex()).setStart();
     }
 
     /**
      * Validate the arguments and payloads from a commandMap generated by CommandParser.
-     * User input is valid if no exceptions are thrown.
-     *
-     * @param arguments Argument-Payload map generated by CommandParser
-     * @throws BadCommandException If the arguments have any issues
+     * User input is valid if no arguments are thrown.
      */
-    @Override
     public void validateCommand(HashMap<String, String> arguments) throws BadCommandException {
         if (arguments.size() != COMMAND_NUM_OF_ARGUMENTS) {
             throw new BadCommandException(COMMAND_INVALID_ARGUMENTS_MESSAGE);
