+++ src/main/java/wellnus/focus/command/StopCommand.java
@@ -1,33 +1,32 @@
-package wellnus.focus;
+package wellnus.focus.command;
 
 import wellnus.command.Command;
 import wellnus.exception.BadCommandException;
+import wellnus.focus.feature.FocusManager;
+import wellnus.focus.feature.Session;
 import wellnus.ui.TextUi;
 
 import java.util.HashMap;
 
-/**
- * Represents a command to check the time left in the current session.
- */
-public class CheckCommand extends Command {
-
-    private static final String COMMAND_KEYWORD = "check";
+public class StopCommand extends Command {
+    private static final String COMMAND_KEYWORD = "stop";
     private static final int COMMAND_NUM_OF_ARGUMENTS = 1;
-    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'check'";
+    private static final String COMMAND_INVALID_ARGUMENTS_MESSAGE = "Invalid command, expected 'stop'";
     private static final String NO_ADDITIONAL_MESSAGE = "";
-    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be check.";
-    private final String CHECK_OUTPUT = "Time left: ";
+    private static final String COMMAND_KEYWORD_ASSERTION = "The key should be stop.";
+    private static final String STOP_MESSAGE = "Your focus session has ended.";
+
     private final Session session;
     private final TextUi textUi;
 
     /**
-     * Constructs a CheckCommand object.
+     * Constructs a StopCommand object.
      * Session in FocusManager is passed into this class.
      *
      * @param arguments Argument-Payload Hashmap generated by CommandParser
      * @param session   Session object which is an arraylist of Countdowns
      */
-    public CheckCommand(HashMap<String, String> arguments, Session session) {
+    public StopCommand(HashMap<String, String> arguments, Session session) {
         super(arguments);
         this.session = session;
         this.textUi = new TextUi();
@@ -56,9 +55,11 @@ public class CheckCommand extends Command {
         return FocusManager.FEATURE_NAME;
     }
 
+
     /**
-     * Checks the current time left in the current countdown.
-     * Prints the time left in the current countdown.
+     * Prints message to indicate session has ended.
+     * Stops the session.
+     * Resets the current countdown index.
      */
     @Override
     public void execute() {
@@ -68,9 +69,9 @@ public class CheckCommand extends Command {
             textUi.printErrorFor(badCommandException, NO_ADDITIONAL_MESSAGE);
             return;
         }
-        int minutes = session.getSession().get(session.getCurrentCountdownIndex()).getMinutes();
-        int seconds = session.getSession().get(session.getCurrentCountdownIndex()).getSeconds();
-        textUi.printOutputMessage(CHECK_OUTPUT + String.format("%d:%d", minutes, seconds));
+        textUi.printOutputMessage(STOP_MESSAGE);
+        session.getSession().get(session.getCurrentCountdownIndex()).setStop();
+        session.resetCurrentCountdownIndex();
     }
 
     /**
